% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqaf_metrics.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{Calculate and print the overall proportion of the provided metric on the records we loaded above}
\usage{
calculate_proportion(
  death_records,
  metric,
  metric_description,
  print_output = TRUE
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{metric}{character string corresponding to the boolean column in death_records we want to calculate proportion for}

\item{metric_description}{string description for print output}

\item{print_output}{boolean, should output be printed to the console? Default TRUE}
}
\value{
number, proportion of records with metric as TRUE
}
\description{
Calculate and print the overall proportion of the provided metric on the records we loaded above
}
\keyword{internal}
