% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqaf_metrics.R
\name{calculate_proportion_by_column}
\alias{calculate_proportion_by_column}
\title{Calculate the proportion of the provided metric for each distinct value in supplied column}
\usage{
calculate_proportion_by_column(
  death_records,
  metric,
  column,
  print_output = TRUE,
  num_print = NA
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{metric}{character string corresponding to the boolean column in death_records we want to calculate proportion for}

\item{column}{string of column name we want to group by}

\item{print_output}{boolean, should output be printed to the console? Default TRUE}

\item{num_print}{number of groups in column we would like to be printed to the console. Default NA}
}
\value{
dataframe with one column corresponding to the unique groups in "column" and another column corresponding to the proportions of records with metric as TRUE for each group
}
\description{
Calculate the proportion of the provided metric for each distinct value in supplied column
}
\keyword{internal}
