% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_not_certified_within_required_period.R
\name{certifier_proportion_not_certified_within_required_period}
\alias{certifier_proportion_not_certified_within_required_period}
\title{Calculate proportion of records not certified within required period by certifier}
\usage{
certifier_proportion_not_certified_within_required_period(
  death_records,
  date_of_death_column,
  date_certified_column,
  number_of_days,
  certifier_name_column
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{date_of_death_column}{string, date of death column in death_records}

\item{date_certified_column}{string, date certified column in death_records}

\item{number_of_days}{number, days that a record must be certified by}

\item{certifier_name_column}{string, certifier name column in death_records}
}
\value{
dataframe with one column corresponding to the unique certifiers by name and another column corresponding to the proportions of records with date certified not within required days
}
\description{
Calculate proportion of records not certified within required period by certifier
}
\examples{
certifier_prop <- certifier_proportion_not_certified_within_required_period(
  synthetic_death_records,
  date_of_death_column = "Date of Death",
  date_certified_column = "Date Certified",
  number_of_days = 5,
  certifier_name_column = "Certifier Name"
)

}
