% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_with_incomplete_funeral_director_fields.R
\name{certifier_proportion_with_incomplete_funeral_director_fields}
\alias{certifier_proportion_with_incomplete_funeral_director_fields}
\title{Calculate proportion of records with at least one incomplete funeral director field by certifier}
\usage{
certifier_proportion_with_incomplete_funeral_director_fields(
  death_records,
  funeral_director_columns,
  unknown_responses = c(),
  age_column = "Age",
  marital_status_column = NULL,
  age_marital_low = 12,
  occupation_column = NULL,
  age_occupation_low = 14,
  industry_column = NULL,
  age_industry_low = 14,
  armed_forces_column = NULL,
  age_armed_low = 14,
  certifier_name_column,
  number_certifier_proportions = 3
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{funeral_director_columns}{character vector of column names in death_records for fields filled out by funeral directors}

\item{unknown_responses}{character vector of strings corresponding to "unknown" in funeral director fields. Default to empty vector.}

\item{age_column}{string, age column in death_records}

\item{marital_status_column}{string, marital status column in death_records. NULL if not in death_records. Default NULL.}

\item{age_marital_low}{number, low cutoff for marital status age. Default 10.}

\item{occupation_column}{string, occupation column in death_records. NULL if not in death_records. Default NULL.}

\item{age_occupation_low}{number, low cutoff for occupation age. Default 14.}

\item{industry_column}{string, industry column in death_records. NULL if not in death_records. Default NULL.}

\item{age_industry_low}{number, low cutoff for industry age. Default 14.}

\item{armed_forces_column}{string, armed forces column in death_records. NULL if not in death_records. Default NULL.}

\item{age_armed_low}{number, low cutoff for armed forces age. Default 14.}

\item{certifier_name_column}{string, certifier name column in death_records}

\item{number_certifier_proportions}{number of certifier proportions to display. Default 3.}
}
\value{
dataframe with one column corresponding to the unique certifiers by name and another column corresponding to the proportions of records with at least one incomplete funeral director field
}
\description{
Calculate proportion of records with at least one incomplete funeral director field by certifier
}
\examples{
certifier_prop <- certifier_proportion_with_incomplete_funeral_director_fields(
  synthetic_death_records,
  funeral_director_columns = c(
    "Decdent's Legal Name",
    "Sex",
    "Social Security Number",
    "Age",
    "Under 1 Year",
    "Under 1 Day",
    "Date of Birth",
    "Birthplace",
    "Residence-State",
    "County",
    "City or Town",
    "Street and Number",
    "Apt. No.",
    "ZIP Code",
    "Inside City Limits",
    "Ever in US Armed Forces",
    "Marital Status at Time of Death",
    "Surviving Spouse's Name",
    "Father's Name",
    "Mother's Name Prior to First Marriage",
    "Informant's Name",
    "Relationship to Decedent",
    "Mailing Address",
    "Place of Death",
    "Facility Name",
    "Facility City, State, and ZIP Code",
    "County of Death",
    "Method of Disposition",
    "Place of Disposition",
    "Location - City, Town, and State",
    "Name and Complete Address of Funeral Facility",
    "Signature of Funeral Service Licensee or Other Agent",
    "License Number",
    "Decedent's Education",
    "Race White", "Race Black or African American",
    "Race American Indian or Alaska Native",
    "Race Asian Indian",
    "Race Chinese",
    "Race Filipino",
    "Race Japanese",
    "Race Korean",
    "Race Vietnamese",
    "Race Other Asian",
    "Race Native Hawaiian",
    "Race Guamanian or Chamorro",
    "Race Samoan",
    "Race Other Pacific Islander",
    "Race Other",
    "Hispanic No",
    "Hispanic Mexican",
    "Hispanic Puerto Rican",
    "Hispanic Cuban",
    "Hispanic Other",
    "Decedent's Usual Occupation",
    "Kind of Business/Industry",
    "Funeral Facility"
  ),
  unknown_responses = c("Unknown", "UNK"),
  age_column = "Age",
  marital_status_column = NULL,
  age_marital_low = 10,
  occupation_column = NULL,
  age_occupation_low = 14,
  industry_column = NULL,
  age_industry_low = 14,
  armed_forces_column = NULL,
  age_armed_low = 14,
  certifier_name_column = "Certifier Name",
  number_certifier_proportions = 3
)


}
