% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_with_incomplete_medical_certifier_fields.R
\name{certifier_proportion_with_incomplete_medical_certifier_fields}
\alias{certifier_proportion_with_incomplete_medical_certifier_fields}
\title{Calculate proportion of records with at least one incomplete medical certifier field by certifier}
\usage{
certifier_proportion_with_incomplete_medical_certifier_fields(
  death_records,
  medical_columns,
  unknown_responses = c(),
  sex_column = "Sex",
  pregnancy_column = "Pregnancy Status",
  age_column = "Age",
  age_pregnancy_low = 5,
  age_pregnancy_high = 74,
  certifier_name_column = "Certifier Name",
  number_certifier_proportions = 3
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{medical_columns}{character vector of column names in death_records for fields filled out by medical certifiers}

\item{unknown_responses}{character vector of strings corresponding to "unknown" in funeral director fields. Default to empty vector.}

\item{sex_column}{string, sex column in death_records}

\item{pregnancy_column}{string, pregnancy column in death_records}

\item{age_column}{string, age column in death_records}

\item{age_pregnancy_low}{number, low cutoff for pregnancy age}

\item{age_pregnancy_high}{number, high cutoff for pregnancy age}

\item{certifier_name_column}{string, certifier name column in death_records}

\item{number_certifier_proportions}{number of certifier proportions to display. Default 3.}
}
\value{
dataframe with one column corresponding to the unique certifiers by name and another column corresponding to the proportions of records with at least one incomplete medical certifier field
}
\description{
Calculate proportion of records with at least one incomplete medical certifier field by certifier
}
