% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_with_one_cause.R
\name{certifier_proportion_with_one_cause}
\alias{certifier_proportion_with_one_cause}
\title{Calculate proportion of records not certified within required period by certifier}
\usage{
certifier_proportion_with_one_cause(
  death_records,
  cause_of_death_columns,
  certifier_name_column = "Certifier Name"
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{cause_of_death_columns}{character vector, cause of death columns in death_records}

\item{certifier_name_column}{string, certifier name column in death_records}
}
\value{
dataframe with one column corresponding to the unique certifiers by name and another column corresponding to the proportions of records with one cause of death
}
\description{
Calculate proportion of records not certified within required period by certifier
}
