% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_with_incomplete_funeral_director_fields.R
\name{check_funeral_columns}
\alias{check_funeral_columns}
\title{Check which funeral director columns are in death records data}
\usage{
check_funeral_columns(death_records, funeral_director_columns)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{funeral_director_columns}{character vector of column names in death_records for fields filled out by funeral directors}
}
\value{
funeral_director_columns subset to those appearing in death_records
}
\description{
Check which funeral director columns are in death records data
}
\keyword{internal}
