% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_with_incomplete_medical_certifier_fields.R
\name{check_medical_columns}
\alias{check_medical_columns}
\title{Check which medical certifier columns are in death records data}
\usage{
check_medical_columns(death_records, medical_columns)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{medical_columns}{character vector of column names in death_records for fields filled out by medical certifiers}
}
\value{
medical_columns subset to those appearing in death_records
}
\description{
Check which medical certifier columns are in death records data
}
\keyword{internal}
