% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqaf_metrics.R
\name{load_death_records}
\alias{load_death_records}
\title{Use supplied path to death records and load as a dataframe}
\usage{
load_death_records(file)
}
\arguments{
\item{file}{file path to death records CSV, with rows corresponding to
records and columns corresponding to record attributes}
}
\value{
data frame with death records information
}
\description{
Use supplied path to death records and load as a dataframe
}
