% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_with_incomplete_funeral_director_fields.R
\name{parse_incomplete_funeral_director_fields}
\alias{parse_incomplete_funeral_director_fields}
\title{Add column for incomplete funeral director fields}
\usage{
parse_incomplete_funeral_director_fields(
  death_records,
  funeral_director_columns,
  unknown_responses,
  age_cutoff_colummns,
  age_column,
  marital_status_column,
  age_marital_low,
  occupation_column,
  age_occupation_low,
  industry_column,
  age_industry_low,
  armed_forces_column,
  age_armed_low
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{funeral_director_columns}{character vector of column names in death_records for fields filled out by funeral directors}

\item{unknown_responses}{character vector of strings corresponding to "unknown" in funeral director fields. Default to empty vector.}

\item{age_cutoff_colummns}{character vector of column names in death_records corresponding to columns requiring age cutoffs (such as marital status, etc.)}

\item{age_column}{string, age column in death_records}

\item{marital_status_column}{string, marital status column in death_records. NULL if not in death_records.}

\item{age_marital_low}{number, low cutoff for marital status age}

\item{occupation_column}{string, occupation column in death_records. NULL if not in death_records.}

\item{age_occupation_low}{number, low cutoff for occupation age}

\item{industry_column}{string, industry column in death_records. NULL if not in death_records.}

\item{age_industry_low}{number, low cutoff for industry age}

\item{armed_forces_column}{string, armed forces column in death_records. NULL if not in death_records.}

\item{age_armed_low}{number, low cutoff for armed forces age}
}
\value{
death_records with an additional boolean column "Incomplete Funeral Director Fields", specifying if the record has at least one incomplete funeral director field
}
\description{
Add column for incomplete funeral director fields
}
\keyword{internal}
