% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_with_incomplete_medical_certifier_fields.R
\name{parse_incomplete_medical_certifier_fields}
\alias{parse_incomplete_medical_certifier_fields}
\title{Add column for incomplete medical certifier fields}
\usage{
parse_incomplete_medical_certifier_fields(
  death_records,
  medical_columns,
  unknown_responses,
  sex_column,
  pregnancy_column,
  age_column,
  age_pregnancy_low,
  age_pregnancy_high
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{medical_columns}{character vector of column names in death_records for fields filled out by medical certifiers}

\item{unknown_responses}{character vector of strings corresponding to "unknown" in funeral director fields. Default to empty vector.}

\item{sex_column}{string, sex column in death_records}

\item{pregnancy_column}{string, pregnancy column in death_records}

\item{age_column}{string, age column in death_records}

\item{age_pregnancy_low}{number, low cutoff for pregnancy age}

\item{age_pregnancy_high}{number, high cutoff for pregnancy age}
}
\value{
death_records with an additional boolean column "Incomplete Medical Certifier Fields", specifying if the record has at least one incomplete medical certifier field
}
\description{
Add column for incomplete medical certifier fields
}
\keyword{internal}
