% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_not_certified_within_required_period.R
\name{parse_not_certified_within_required_period}
\alias{parse_not_certified_within_required_period}
\title{Add column for not certified within required period}
\usage{
parse_not_certified_within_required_period(
  death_records,
  date_of_death_column = "Date of Death",
  date_certified_column = "Date Certified",
  number_of_days = 5
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{date_of_death_column}{string, date of death column in death_records}

\item{date_certified_column}{string, date certified column in death_records}

\item{number_of_days}{number, days that a record must be certified by}
}
\value{
death_records with an additional boolean column "Not Within Required Days", specifying if the record was not certified within the specified period
}
\description{
Add column for not certified within required period
}
\keyword{internal}
