% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_with_one_cause.R
\name{parse_with_one_cause}
\alias{parse_with_one_cause}
\title{Add column for one cause of death}
\usage{
parse_with_one_cause(death_records, cause_of_death_columns)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{cause_of_death_columns}{character vector, cause of death columns in death_records}
}
\value{
death_records with an additional boolean column "Single Record Axis COD", specifying if the record has only one cause of death specified
}
\description{
Add column for one cause of death
}
\keyword{internal}
