% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_with_unsuitable_underlying.R
\name{parse_with_unsuitable_underlying}
\alias{parse_with_unsuitable_underlying}
\title{Add column for unsuitable underlying causes of death}
\usage{
parse_with_unsuitable_underlying(
  death_records,
  underlying_cause_of_death_column
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{underlying_cause_of_death_column}{string, underlying cause of death column in death_records}
}
\value{
death_records with an additional boolean column "Unsuitable Underlying", specifying if the record was not certified within the specified period
}
\description{
Add column for unsuitable underlying causes of death
}
\keyword{internal}
