% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_not_certified_within_required_period.R
\name{proportion_not_certified_within_required_period}
\alias{proportion_not_certified_within_required_period}
\title{Calculate proportion of records not certified within required period}
\usage{
proportion_not_certified_within_required_period(
  death_records,
  date_of_death_column,
  date_certified_column,
  number_of_days
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{date_of_death_column}{string, date of death column in death_records}

\item{date_certified_column}{string, date certified column in death_records}

\item{number_of_days}{number, days that a record must be certified by}
}
\value{
number, proportion of records not certified within required period
}
\description{
Calculate proportion of records not certified within required period
}
\examples{
prop <- proportion_not_certified_within_required_period(
  synthetic_death_records,
  date_of_death_column = "Date of Death",
  date_certified_column = "Date Certified",
  number_of_days = 5
)

}
