% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_with_incomplete_funeral_director_fields.R
\name{proportion_with_incomplete_funeral_director_fields}
\alias{proportion_with_incomplete_funeral_director_fields}
\title{Calculate proportion of records with at least one incomplete funeral director field}
\usage{
proportion_with_incomplete_funeral_director_fields(
  death_records,
  funeral_director_columns,
  unknown_responses = c(),
  age_column = "Age",
  marital_status_column = NULL,
  age_marital_low = 10,
  occupation_column = NULL,
  age_occupation_low = 14,
  industry_column = NULL,
  age_industry_low = 14,
  armed_forces_column = NULL,
  age_armed_low = 14
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{funeral_director_columns}{character vector of column names in death_records for fields filled out by funeral directors}

\item{unknown_responses}{character vector of strings corresponding to "unknown" in funeral director fields. Default to empty vector.}

\item{age_column}{string, age column in death_records}

\item{marital_status_column}{string, marital status column in death_records. NULL if not in death_records. Default NULL.}

\item{age_marital_low}{number, low cutoff for marital status age. Default 10.}

\item{occupation_column}{string, occupation column in death_records. NULL if not in death_records. Default NULL.}

\item{age_occupation_low}{number, low cutoff for occupation age. Default 14.}

\item{industry_column}{string, industry column in death_records. NULL if not in death_records. Default NULL.}

\item{age_industry_low}{number, low cutoff for industry age. Default 14.}

\item{armed_forces_column}{string, armed forces column in death_records. NULL if not in death_records. Default NULL.}

\item{age_armed_low}{number, low cutoff for armed forces age. Default 14.}
}
\value{
number, proportion of records with at least one incomplete funeral director field
}
\description{
Calculate proportion of records with at least one incomplete funeral director field
}
