% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_with_unsuitable_underlying.R
\name{proportion_with_unsuitable_underlying}
\alias{proportion_with_unsuitable_underlying}
\title{Calculate proportion of records with unsuitable underlying causes of death}
\usage{
proportion_with_unsuitable_underlying(
  death_records,
  underlying_cause_of_death_column
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{underlying_cause_of_death_column}{string, underlying cause of death column in death_records}
}
\value{
number, proportion of records with unsuitable underlying causes of death
}
\description{
Calculate proportion of records with unsuitable underlying causes of death
}
\examples{
prop <- proportion_with_unsuitable_underlying(
  synthetic_death_records,
  underlying_cause_of_death_column = "Underlying COD"
)

}
