% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_with_incomplete_demographic.R
\name{subset_records_for_cutoffs}
\alias{subset_records_for_cutoffs}
\title{Subset death records data to relevant records based on age cutoffs}
\usage{
subset_records_for_cutoffs(
  death_records,
  mc,
  sex_column,
  pregnancy_column,
  age_column,
  age_pregnancy_low,
  age_pregnancy_high,
  marital_status_column,
  age_marital_low,
  occupation_column,
  age_occupation_low,
  industry_column,
  age_industry_low,
  armed_forces_column,
  age_armed_low
)
}
\arguments{
\item{death_records}{death records dataframe with rows corresponding to
records and columns corresponding to record attributes}

\item{mc}{string, column we are calculating proportions for}

\item{sex_column}{string, sex column in death_records}

\item{pregnancy_column}{string, pregnancy column in death_records}

\item{age_column}{string, age column in death_records}

\item{age_pregnancy_low}{number, low cutoff for pregnancy age}

\item{age_pregnancy_high}{number, high cutoff for pregnancy age}

\item{marital_status_column}{string, marital status column in death_records. NULL if not in death_records.}

\item{age_marital_low}{number, low cutoff for marital status age}

\item{occupation_column}{string, occupation column in death_records. NULL if not in death_records.}

\item{age_occupation_low}{number, low cutoff for occupation age}

\item{industry_column}{string, industry column in death_records. NULL if not in death_records.}

\item{age_industry_low}{number, low cutoff for industry age}

\item{armed_forces_column}{string, armed forces column in death_records. NULL if not in death_records.}

\item{age_armed_low}{number, low cutoff for armed forces age}
}
\value{
death_records, subset to the relevant records in mc based on cutoffs
}
\description{
Subset death records data to relevant records based on age cutoffs
}
