% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_death_records}
\alias{synthetic_death_records}
\title{Synthetic Death Records}
\format{
An object of class \code{data.frame} with 4000 rows and 71 columns.
}
\usage{
synthetic_death_records
}
\description{
An example of data used for metric calculation, which should have a format where each row corresponds to one death record and each column corresponds to an attribute for the death record, such as "Date Certified" or "Underlying Cause of Death". ICD codes included in the data should not include periods (".").
}
\keyword{datasets}
