% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqaf_metrics.R
\name{unspecified_arguments_error}
\alias{unspecified_arguments_error}
\title{Check if not all arguments are specified}
\usage{
unspecified_arguments_error(arg_passed, func)
}
\arguments{
\item{arg_passed}{character vector, arguments passed to function}

\item{metric}{character string corresponding to the boolean column in death_records we want to calculate proportion for}
}
\value{
nothing, will error if not all arguments are specified
}
\description{
Check if not all arguments are specified
}
\keyword{internal}
