% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unsuitable_COD_codes}
\alias{unsuitable_COD_codes}
\title{Unsuitable Cause of Death Codes}
\format{
An object of class \code{data.frame} with 389 rows and 3 columns.
}
\source{
Flagg, 2021 (https://stacks.cdc.gov/view/cdc/100414).
}
\usage{
unsuitable_COD_codes
}
\description{
A data file (unsuitable_COD_codes.csv), located in the data directory, containing ICD10 codes that have been identified as unsuitable underlying causes of death as listed in (Flagg 2021).
}
\details{
Note that the file contains both specific four digit ICD-10 codes, which don't include the decimal point (e.g., I959, referring to "Hypotension, unspecified") and three digit ICD-10 codes for broader categories, e.g., J18, referring to "Pneumonia, organism unspecified".

Three digit codes imply the inclusion of all related four digit codes, e.g., J18, referring to "Pneumonia, organism unspecified", also implicitly means the inclusion of J180, referring to "Bronchopneumonia, unspecified", J181, referring to "Lobar pneumonia, unspecified", etc.
}
\keyword{datasets}
